Refractor data datatypes (move to separate .d.ts file)
