+++ apps/astro-app/src/components/data/form-helper.ts
@@ -2,7 +2,7 @@
  * Helpers for form creation. Does anything that may universally apply to all forms of data entry.
  */
 import { zodResolver } from "@hookform/resolvers/zod";
-import type { DataColumn, TableInfo } from "@/env";
+import type { DataColumn, TableInfo } from "@/types/data";
 import { toSnakeCase } from "@/utils";
 import { getFallbackValue, getZodType } from "@/utils/data";
 import { useForm, type FieldErrors, type UseFormReturn } from "react-hook-form";
