+++ apps/astro-app/src/pages/data/[databaseName]/[tableName]/dataentry.astro
@@ -4,7 +4,7 @@
  *
  * @TODO redirect if dataHarvesting is disabled
  */
-import type { DatabaseInfo, TableInfo } from "@/env";
+import type { DatabaseInfo, TableInfo } from "@/types/data";
 import config from "@universal_root/config.yml";
 import Layout from "@/layouts/Layout.astro";
