+++ apps/astro-app/src/pages/data/[databaseName]/[tableName]/entries/[...slug].astro
@@ -2,7 +2,7 @@
 /*
  * @TODO redirect if dataHarvesting is disabled
  */
-import type { DatabaseInfo, DescriptorInfo, TableInfo } from "@/env";
+import type { DatabaseInfo, DescriptorInfo, TableInfo } from "@/types/data";
 import config from "@universal_root/config.yml";
 import Layout from "@/layouts/Layout.astro";
